refactor: reduce deepEqual calls by using diff store
