+++ ui/src/entities/keys/lib/index.ts
@@ -1,3 +1,3 @@
 export * from './build-template'
 export * from './constants'
-export * from './key-name'
+export * from './keys'
